+++ src/webapp/src/components/analyze/Design.vue
@@ -434,7 +434,15 @@ export default {
             <span class="has-text-weight-bold">{{ design.label }}</span>
           </h3>
         </div>
-        <p v-if="design.description">{{ design.description }}</p>
+        <p
+          v-if="
+            design.description &&
+              design.description.length &&
+              design.description != design.label
+          "
+        >
+          {{ design.description }}
+        </p>
       </div>
 
       <div class="column">